chatbot-service/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
requests 
python-dotenv
langdetect
gtts
